重构野生宠物巡游和Talk Bubble的部分代码
NpcInfo 里把巡游配置收进了 NpcWanderInfo，气泡配置收进了 NpcWanderBubbleInfo：
NpcInfo.cs (Assets/Scripts/Scene Specific/Map/Entities/Npc/NpcInfo.cs)
巡游逻辑从 NpcHandler.cs 拆到独立脚本：
MapWildNpcWanderController.cs (Assets/Scripts/Scene Specific/Map/Entities/Npc/MapWildNpcWanderController.cs)
气泡逻辑保留为独立控制器：
MapWildNpcSpriteBubbleController.cs(D:/BBB_projects/Seer2-Restart-Lite/Assets/Scripts/Scene Specific/Map/Entities/Npc/MapWildNpcSpriteBubbleController.cs)
141.xml 从扁平属性改成：
<wander ...><bubble ...><self>...</self></bubble></wander>
141.xml (Assets/Resources/Data/Maps/141.xml)
