+++ Assets/Resources/Data/Maps/141.xml
@@ -7,12 +7,14 @@
 	</music>
 	<initialPoint>550,100</initialPoint>
 	<entities>
-		<npc id="14101" resId="Pets/pet/18.png" name="Lv8 小树根" size="43,40" pos="220,210" namePos="0,0" wander="true" wanderRadius="220" wanderSpeed="34" wanderIdle="2,3" wanderStep="12,26" wanderMoveTick="0.05" wanderSnap="1" wanderBob="1.25" wanderBobSpeed="2" wanderOriginalFacesRight="true" wanderBubbleChance="0.3" wanderBubbleDuration="2.5">
-			<wanderBubble>
-				<self>有时我会想，是不是要一辈子呆在幻影崖呢？</self>
-				<self>等我长大之后，一定不会再被布鲁托欺负了！</self>
-				<self></self>
-			</wanderBubble>
+		<npc id="14101" resId="Pets/pet/18.png" name="Lv8 小树根" size="43,40" pos="220,210" namePos="0,0">
+			<wander enabled="true" radius="220" speed="34" idle="2,3" step="12,26" moveTick="0.05" snap="1" bob="1.25" bobSpeed="2" originalFacesRight="true">
+				<bubble chance="0.3" duration="2.5">
+					<self>有时我会想，是不是要一辈子呆在幻影崖呢？</self>
+					<self>等我长大之后，一定不会再被布鲁托欺负了！</self>
+					<self></self>
+				</bubble>
+			</wander>
 			<battle>
 				<branch id="default">
 					<settings count="6" capture="true"/>
@@ -47,15 +49,17 @@
 				</button>
 			</eventHandler>
 		</npc>
-		<npc id="14102" resId="Pets/pet/18.png" name="Lv9 小树根" size="43,40" pos="380,70" namePos="0,0" wander="true" wanderRadius="220" wanderSpeed="34" wanderIdle="2,3" wanderStep="12,26" wanderMoveTick="0.05" wanderSnap="1" wanderBob="1.25" wanderBobSpeed="2" wanderOriginalFacesRight="true" wanderBubbleChance="0.30" wanderBubbleDuration="2.5">
-			<wanderBubble>
-				<self>严肃光合作用！</self>
-				<self>兄弟你别盯着我看。</self>
-				<self>猛猛扎根ing！</self>
-				<self>这里有点安静。。。</self>
-				<self>今天适合扎根~</self>
-				<self>这里怎么有一个低——等机器人~</self>
-			</wanderBubble>
+		<npc id="14102" resId="Pets/pet/18.png" name="Lv9 小树根" size="43,40" pos="380,70" namePos="0,0">
+			<wander enabled="true" radius="220" speed="34" idle="2,3" step="12,26" moveTick="0.05" snap="1" bob="1.25" bobSpeed="2" originalFacesRight="true">
+				<bubble chance="0.30" duration="2.5">
+					<self>严肃光合作用！</self>
+					<self>兄弟你别盯着我看。</self>
+					<self>猛猛扎根ing！</self>
+					<self>这里有点安静。。。</self>
+					<self>今天适合扎根~</self>
+					<self>这里怎么有一个低——等机器人~</self>
+				</bubble>
+			</wander>
 			<battle>
 				<branch id="default">
 					<settings count="6" capture="true"/>
