+++ Assets/Scripts/Scene Specific/Map/Entities/Npc/MapWildNpcSpriteBubbleController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1acb72fba60466d45be30691cb2ab092
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
