+++ Assets/Scripts/Scene Specific/Map/Entities/Npc/MapWildNpcWanderController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 679d6a58ff14347418ca5b109db5d42b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
